+++ src/components/Cart/CartItem/CartItem.styles.ts
