+++ src/components/Product/ProductItem/ProductItem.stories.tsx
@@ -12,7 +12,7 @@ const mock: Product = {
 
 const productItem = {
   component: ProductItem,
-  title: 'Cart/ProductItem',
+  title: 'Product/ProductItem',
   tags: ['autodocs'],
   args: {
     product: mock,
